project1/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.196
Pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
matplotlib==3.7.2